Correct an error in the envar sum test (#92)
Your example tests routinely demonstrate a left-to-right precedence
ordering when setting the envars. However, one test mistakenly computes
the target sum using right-to-left precedence. I have corrected it in
this PR and provided detailed comments for each test explaining how the
sum is being calculated. This should hopefully help to avoid such
problems in the future.

Signed-off-by: Ralph Castain <rhc@pmix.org>